Feature: Add setting to toggle link reference to optionally be the full note name (#30)
* Update tagSearch.ts to allow full link name (#26)

Allows full link name to be shown instead of *

* feat: add option to display full link name as reference

This commit introduces a new feature where users can choose to display the full link title as a reference at the end of the line in the Obsidian Tag Page plugin. By default, the referenced content will end with a link aliased as `*`. However, when the new setting is toggled on, the full file name will be used for the link. This provides users with more flexibility in how their tag pages are formatted. The README has been updated to reflect this new feature.

---------

Co-authored-by: duckfarts <24978179+duckfarts@users.noreply.github.com>